+++ common/scala/src/main/scala/org/apache/openwhisk/core/entity/WhiskActivation.scala
@@ -173,13 +173,13 @@ object WhiskActivation
   private val filtersDdoc = dbConfig.activationsFilterDdoc
 
   /** The main view for activations, keyed by namespace, sorted by date. */
-  override lazy val view = WhiskEntityQueries.view(mainDdoc, collectionName)
+  override lazy val view = WhiskQueries.view(mainDdoc, collectionName)
 
   /**
    * A view for activations in a namespace additionally keyed by action name
    * (and package name if present) sorted by date.
    */
-  lazy val filtersView = WhiskEntityQueries.view(filtersDdoc, collectionName)
+  lazy val filtersView = WhiskQueries.view(filtersDdoc, collectionName)
 
   override implicit val serdes = jsonFormat13(WhiskActivation.apply)
 
@@ -204,7 +204,7 @@ object WhiskActivation
                                   upto: Option[Instant] = None,
                                   stale: StaleParameter = StaleParameter.No)(
     implicit transid: TransactionId): Future[Either[List[JsObject], List[WhiskActivation]]] = {
-    import WhiskEntityQueries.TOP
+    import WhiskQueries.TOP
     val convert = if (includeDocs) Some((o: JsObject) => Try { serdes.read(o) }) else None
     val startKey = List(namespace.addPath(path).asString, since map { _.toEpochMilli } getOrElse 0)
     val endKey = List(namespace.addPath(path).asString, upto map { _.toEpochMilli } getOrElse TOP, TOP)
