+++ common/scala/src/main/scala/org/apache/openwhisk/core/entity/WhiskPackage.scala
@@ -202,7 +202,7 @@ object WhiskPackage
 
   override val cacheEnabled = true
 
-  lazy val publicPackagesView: View = WhiskEntityQueries.view(collection = s"$collectionName-public")
+  lazy val publicPackagesView: View = WhiskQueries.view(collection = s"$collectionName-public")
 }
 
 /**
